refactor(ic-certified-assets): remove propose/commit-proposed workflow
The propose -> compute_evidence -> commit_proposed_batch flow was carried
over from the legacy implementation and is no longer used by the plugin.
Removes the four canister methods, their argument types, the batch-side
state they relied on, and renames the now state-hash-only evidence module.
