+++ ic-certified-assets/src/tests.rs
@@ -5,10 +5,9 @@ use crate::state_machine::{ComputationStatus, StableStateV2, State, BATCH_EXPIRY
 use crate::system_context::canister_env::CanisterEnv;
 use crate::system_context::SystemContext;
 use crate::types::{
-    AssetProperties, BatchId, BatchOperation, CommitBatchArguments, CommitProposedBatchArguments,
-    ComputeEvidenceArguments, CreateAssetArguments, CreateChunkArg, DeleteAssetArguments,
-    DeleteBatchArguments, GetArg, GetChunkArg, ListRequest, SetAssetContentArguments,
-    SetAssetPropertiesArguments,
+    AssetProperties, BatchId, BatchOperation, CommitBatchArguments, CreateAssetArguments,
+    CreateChunkArg, DeleteAssetArguments, DeleteBatchArguments, GetArg, GetChunkArg, ListRequest,
+    SetAssetContentArguments, SetAssetPropertiesArguments,
 };
 use crate::url::{url_decode, url_encode, UrlDecodeError};
 use crate::CreateChunksArg;
@@ -263,50 +262,6 @@ fn create_assets(
     batch_id
 }
 
-fn create_assets_by_proposal(
-    state: &mut State,
-    system_context: &SystemContext,
-    assets: Vec<AssetBuilder>,
-) -> BatchId {
-    let batch_id = state.create_batch(system_context).unwrap();
-
-    let operations = assemble_create_assets_and_set_contents_operations(
-        state,
-        system_context,
-        assets,
-        &batch_id,
-    );
-
-    state
-        .propose_commit_batch(CommitBatchArguments {
-            batch_id: batch_id.clone(),
-            operations,
-        })
-        .unwrap();
-
-    let evidence = run_computation_until_completion(|_progress| {
-        state.compute_evidence(&ComputeEvidenceArguments {
-            batch_id: batch_id.clone(),
-            max_iterations: Some(100),
-        })
-    })
-    .unwrap();
-
-    run_computation_until_completion(|progress| {
-        state.commit_proposed_batch(
-            &CommitProposedBatchArguments {
-                batch_id: batch_id.clone(),
-                evidence: evidence.clone(),
-            },
-            progress,
-            system_context,
-        )
-    })
-    .unwrap();
-
-    batch_id
-}
-
 fn assemble_create_assets_and_set_contents_operations(
     state: &mut State,
     system_context: &SystemContext,
@@ -358,12 +313,6 @@ fn assemble_create_assets_and_set_contents_operations(
         }
     }
     operations
-}
-
-fn delete_batch(state: &mut State, batch_id: BatchId) {
-    state
-        .delete_batch(DeleteBatchArguments { batch_id })
-        .unwrap();
 }
 
 fn lookup_header<'a>(response: &'a HttpResponse, header: &str) -> Option<&'a str> {
@@ -668,47 +617,6 @@ fn serve_fallback_v1() {
 }
 
 #[test]
-fn can_create_assets_using_batch_proposal_api() {
-    let mut state = State::default();
-    let system_context = mock_system_context();
-
-    const BODY: &[u8] = b"<!DOCTYPE html><html></html>";
-
-    let batch_id = create_assets_by_proposal(
-        &mut state,
-        &system_context,
-        vec![AssetBuilder::new("/contents.html", "text/html").with_encoding("identity", vec![BODY])],
-    );
-
-    let response = certified_http_request(
-        &state,
-        RequestBuilder::get("/contents.html")
-            .with_header("Accept-Encoding", "gzip,identity")
-            .build(),
-    );
-
-    assert_eq!(response.status_code, 200);
-    assert_eq!(response.body.as_ref(), BODY);
-
-    // Try to update a completed batch.
-    let error_msg = state
-        .create_chunk(
-            CreateChunkArg {
-                batch_id,
-                content: ByteBuf::new(),
-            },
-            &system_context,
-        )
-        .unwrap_err();
-
-    let expected = "batch not found";
-    assert!(
-        error_msg.contains(expected),
-        "expected '{expected}' error, got: {error_msg}"
-    );
-}
-
-#[test]
 fn batches_are_dropped_after_timeout() {
     let mut state = State::default();
     let mut system_context = mock_system_context();
@@ -741,170 +649,6 @@ fn batches_are_dropped_after_timeout() {
         Err(err) if err.contains("batch not found") => (),
         other => panic!("expected 'batch not found' error, got: {other:?}"),
     }
-}
-
-#[test]
-fn can_propose_commit_batch_exactly_once() {
-    let mut state = State::default();
-    let system_context = mock_system_context();
-
-    let batch_1 = state.create_batch(&system_context).unwrap();
-
-    let args = CommitBatchArguments {
-        batch_id: batch_1.clone(),
-        operations: vec![],
-    };
-    assert_eq!(Ok(()), state.propose_commit_batch(args.clone()));
-    match state.propose_commit_batch(args) {
-        Err(err)
-            if err == format!("batch {batch_1} already has proposed CommitBatchArguments",) => {}
-        other => panic!("expected batch already proposed error, got: {other:?}"),
-    };
-}
-
-#[test]
-fn cannot_create_chunk_in_proposed_batch_() {
-    let mut state = State::default();
-    let system_context = mock_system_context();
-
-    let batch_1 = state.create_batch(&system_context).unwrap();
-
-    let args = CommitBatchArguments {
-        batch_id: batch_1.clone(),
-        operations: vec![],
-    };
-    assert_eq!(Ok(()), state.propose_commit_batch(args));
-
-    const BODY: &[u8] = b"<!DOCTYPE html><html></html>";
-    match state.create_chunk(
-        CreateChunkArg {
-            batch_id: batch_1.clone(),
-            content: ByteBuf::from(BODY.to_vec()),
-        },
-        &system_context,
-    ) {
-        Err(err) if err == format!("batch {batch_1} has been proposed") => {}
-        other => panic!("expected batch already proposed error, got: {other:?}"),
-    }
-    match state.create_chunks(
-        CreateChunksArg {
-            batch_id: batch_1.clone(),
-            content: vec![ByteBuf::from(BODY.to_vec())],
-        },
-        &system_context,
-    ) {
-        Err(err) if err == format!("batch {batch_1} has been proposed") => {}
-        other => panic!("expected batch already proposed error, got: {other:?}"),
-    }
-}
-
-#[test]
-fn batches_with_proposed_commit_args_do_not_expire() {
-    let mut state = State::default();
-    let mut system_context = mock_system_context();
-
-    let batch_1 = state.create_batch(&system_context).unwrap();
-
-    const BODY: &[u8] = b"<!DOCTYPE html><html></html>";
-
-    let _chunk_1 = state
-        .create_chunk(
-            CreateChunkArg {
-                batch_id: batch_1.clone(),
-                content: ByteBuf::from(BODY.to_vec()),
-            },
-            &system_context,
-        )
-        .unwrap();
-
-    let args = CommitBatchArguments {
-        batch_id: batch_1.clone(),
-        operations: vec![],
-    };
-    assert_eq!(Ok(()), state.propose_commit_batch(args));
-
-    system_context.current_timestamp_ns =
-        system_context.current_timestamp_ns + BATCH_EXPIRY_NANOS + 1;
-    let _batch_2 = state.create_batch(&system_context);
-
-    match state.create_chunk(
-        CreateChunkArg {
-            batch_id: batch_1,
-            content: ByteBuf::from(BODY.to_vec()),
-        },
-        &system_context,
-    ) {
-        Err(err) if err.contains("batch not found") => (),
-        other => panic!("expected 'batch not found' error, got: {other:?}"),
-    }
-}
-
-#[test]
-fn batches_with_evidence_do_not_expire() {
-    let mut state = State::default();
-    let mut system_context = mock_system_context();
-
-    let batch_1 = state.create_batch(&system_context).unwrap();
-
-    const BODY: &[u8] = b"<!DOCTYPE html><html></html>";
-
-    let _chunk_1 = state
-        .create_chunk(
-            CreateChunkArg {
-                batch_id: batch_1.clone(),
-                content: ByteBuf::from(BODY.to_vec()),
-            },
-            &system_context,
-        )
-        .unwrap();
-
-    let args = CommitBatchArguments {
-        batch_id: batch_1.clone(),
-        operations: vec![],
-    };
-    assert_eq!(Ok(()), state.propose_commit_batch(args));
-    assert!(run_computation_until_completion(|_progress| {
-        state.compute_evidence(&ComputeEvidenceArguments {
-            batch_id: batch_1.clone(),
-            max_iterations: Some(3),
-        })
-    })
-    .is_ok());
-
-    system_context.current_timestamp_ns =
-        system_context.current_timestamp_ns + BATCH_EXPIRY_NANOS + 1;
-    let _batch_2 = state.create_batch(&system_context);
-
-    match state.create_chunk(
-        CreateChunkArg {
-            batch_id: batch_1.clone(),
-            content: ByteBuf::from(BODY.to_vec()),
-        },
-        &system_context,
-    ) {
-        Err(err) if err == format!("batch {batch_1} has been proposed") => {}
-        other => panic!("expected batch already proposed error, got: {other:?}"),
-    }
-}
-
-#[test]
-fn can_delete_proposed_batch() {
-    let mut state = State::default();
-    let system_context = mock_system_context();
-
-    let batch_1 = state.create_batch(&system_context).unwrap();
-
-    let args = CommitBatchArguments {
-        batch_id: batch_1.clone(),
-        operations: vec![],
-    };
-    assert_eq!(Ok(()), state.propose_commit_batch(args));
-    let delete_args = DeleteBatchArguments { batch_id: batch_1 };
-    assert_eq!(Ok(()), state.delete_batch(delete_args.clone()));
-    assert_eq!(
-        Err("batch not found".to_string()),
-        state.delete_batch(delete_args)
-    );
 }
 
 #[test]
@@ -2489,1621 +2233,6 @@ mod certification_v2 {
 }
 
 #[cfg(test)]
-mod evidence_computation {
-    use std::collections::BTreeMap;
-
-    use super::*;
-    use crate::types::BatchOperation::SetAssetContent;
-    use crate::types::{ClearArguments, ComputeEvidenceArguments, UnsetAssetContentArguments};
-
-    #[test]
-    fn evidence_with_set_single_chunk_asset_content() {
-        let mut state = State::default();
-        let system_context = mock_system_context();
-
-        let batch_1 = state.create_batch(&system_context).unwrap();
-        const BODY: &[u8] = b"<!DOCTYPE html><html></html>";
-        let chunk_1 = state
-            .create_chunk(
-                CreateChunkArg {
-                    batch_id: batch_1.clone(),
-                    content: ByteBuf::from(BODY.to_vec()),
-                },
-                &system_context,
-            )
-            .unwrap();
-
-        let create_asset = CreateAssetArguments {
-            key: "/a/b/c".to_string(),
-            content_type: "text/plain".to_string(),
-            max_age: None,
-            headers: None,
-            enable_aliasing: None,
-            allow_raw_access: None,
-        };
-        let set_asset_content = SetAssetContentArguments {
-            key: "/a/b/c".to_string(),
-            content_encoding: "identity".to_string(),
-            chunk_ids: vec![chunk_1],
-            last_chunk: None,
-            sha256: None,
-        };
-        let cba = CommitBatchArguments {
-            batch_id: batch_1.clone(),
-            operations: vec![
-                BatchOperation::CreateAsset(create_asset),
-                BatchOperation::SetAssetContent(set_asset_content),
-            ],
-        };
-        assert!(state.propose_commit_batch(cba).is_ok());
-        assert!(run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_1.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .is_ok());
-        assert!(run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_1.clone(),
-                max_iterations: Some(1),
-            })
-        })
-        .is_ok());
-    }
-
-    #[test]
-    fn evidence_with_set_multiple_chunk_asset_content() {
-        let mut state = State::default();
-        let system_context = mock_system_context();
-
-        let batch_1 = state.create_batch(&system_context).unwrap();
-        const CHUNK_1_CONTENT: &[u8] = b"<!DOCTYPE html><html></html>";
-        const CHUNK_2_CONTENT: &[u8] = b"there is more content here";
-        let chunk_1 = state
-            .create_chunk(
-                CreateChunkArg {
-                    batch_id: batch_1.clone(),
-                    content: ByteBuf::from(CHUNK_1_CONTENT.to_vec()),
-                },
-                &system_context,
-            )
-            .unwrap();
-        let chunk_2 = state
-            .create_chunk(
-                CreateChunkArg {
-                    batch_id: batch_1.clone(),
-                    content: ByteBuf::from(CHUNK_2_CONTENT.to_vec()),
-                },
-                &system_context,
-            )
-            .unwrap();
-
-        let create_asset = CreateAssetArguments {
-            key: "/a/b/c".to_string(),
-            content_type: "text/plain".to_string(),
-            max_age: None,
-            headers: None,
-            enable_aliasing: None,
-            allow_raw_access: None,
-        };
-        let set_asset_content = SetAssetContentArguments {
-            key: "/a/b/c".to_string(),
-            content_encoding: "identity".to_string(),
-            chunk_ids: vec![chunk_1, chunk_2],
-            last_chunk: None,
-            sha256: None,
-        };
-        let cba = CommitBatchArguments {
-            batch_id: batch_1.clone(),
-            operations: vec![
-                BatchOperation::CreateAsset(create_asset),
-                BatchOperation::SetAssetContent(set_asset_content),
-            ],
-        };
-        assert!(state.propose_commit_batch(cba).is_ok());
-        assert!(run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_1.clone(),
-                max_iterations: Some(4),
-            })
-        })
-        .is_ok());
-        assert!(run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_1.clone(),
-                max_iterations: Some(1),
-            })
-        })
-        .is_ok());
-    }
-
-    #[test]
-    fn evidence_with_create_asset() {
-        let mut state = State::default();
-        let system_context = mock_system_context();
-
-        let batch_id = state.create_batch(&system_context).unwrap();
-        let create_asset = CreateAssetArguments {
-            key: "/a/b/c".to_string(),
-            content_type: "text/plain".to_string(),
-            max_age: None,
-            headers: None,
-            enable_aliasing: None,
-            allow_raw_access: None,
-        };
-        let cba = CommitBatchArguments {
-            batch_id: batch_id.clone(),
-            operations: vec![BatchOperation::CreateAsset(create_asset)],
-        };
-
-        assert!(state.propose_commit_batch(cba).is_ok());
-
-        let compute_args = ComputeEvidenceArguments {
-            batch_id,
-            max_iterations: Some(1),
-        };
-        assert!(run_computation_until_completion(|_progress| {
-            state.compute_evidence(&compute_args)
-        })
-        .is_ok());
-        assert!(run_computation_until_completion(|_progress| {
-            state.compute_evidence(&compute_args)
-        })
-        .is_ok());
-    }
-
-    #[test]
-    fn evidence_with_set_empty_asset_content() {
-        let mut state = State::default();
-        let system_context = mock_system_context();
-
-        let batch_id = state.create_batch(&system_context).unwrap();
-        let create_asset = CreateAssetArguments {
-            key: "/a/b/c".to_string(),
-            content_type: "text/plain".to_string(),
-            max_age: None,
-            headers: None,
-            enable_aliasing: None,
-            allow_raw_access: None,
-        };
-        let set_asset_content = SetAssetContentArguments {
-            key: "/a/b/c".to_string(),
-            content_encoding: "identity".to_string(),
-            chunk_ids: vec![],
-            last_chunk: None,
-            sha256: None,
-        };
-        let cba = CommitBatchArguments {
-            batch_id: batch_id.clone(),
-            operations: vec![
-                BatchOperation::CreateAsset(create_asset),
-                BatchOperation::SetAssetContent(set_asset_content),
-            ],
-        };
-        assert!(state.propose_commit_batch(cba).is_ok());
-
-        assert!(run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_id.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .is_ok());
-        assert!(run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_id.clone(),
-                max_iterations: Some(1),
-            })
-        })
-        .is_ok());
-    }
-
-    #[test]
-    fn evidence_with_no_operations() {
-        let mut state = State::default();
-        let system_context = mock_system_context();
-
-        let batch_id = state.create_batch(&system_context).unwrap();
-        let cba = CommitBatchArguments {
-            batch_id: batch_id.clone(),
-            operations: vec![],
-        };
-        assert!(state.propose_commit_batch(cba).is_ok());
-
-        assert!(run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_id.clone(),
-                max_iterations: Some(1),
-            })
-        })
-        .is_ok());
-    }
-
-    #[test]
-    fn create_asset_same_fields_same_evidence() {
-        let mut state = State::default();
-        let system_context = mock_system_context();
-
-        {
-            let batch_1 = state.create_batch(&system_context).unwrap();
-            assert!(state
-                .propose_commit_batch(CommitBatchArguments {
-                    batch_id: batch_1.clone(),
-                    operations: vec![BatchOperation::CreateAsset(CreateAssetArguments {
-                        key: "/a/b/c".to_string(),
-                        content_type: "".to_string(),
-                        max_age: None,
-                        headers: None,
-                        enable_aliasing: None,
-                        allow_raw_access: None,
-                    }),],
-                })
-                .is_ok());
-            let evidence_1 = run_computation_until_completion(|_progress| {
-                state.compute_evidence(&ComputeEvidenceArguments {
-                    batch_id: batch_1.clone(),
-                    max_iterations: Some(3),
-                })
-            })
-            .unwrap();
-            delete_batch(&mut state, batch_1);
-
-            let batch_2 = state.create_batch(&system_context).unwrap();
-            assert!(state
-                .propose_commit_batch(CommitBatchArguments {
-                    batch_id: batch_2.clone(),
-                    operations: vec![BatchOperation::CreateAsset(CreateAssetArguments {
-                        key: "/a/b/c".to_string(),
-                        content_type: "".to_string(),
-                        max_age: None,
-                        headers: None,
-                        enable_aliasing: None,
-                        allow_raw_access: None,
-                    }),],
-                })
-                .is_ok());
-            let evidence_2 = run_computation_until_completion(|_progress| {
-                state.compute_evidence(&ComputeEvidenceArguments {
-                    batch_id: batch_2.clone(),
-                    max_iterations: Some(3),
-                })
-            })
-            .unwrap();
-            delete_batch(&mut state, batch_2);
-
-            assert_eq!(evidence_1, evidence_2);
-        }
-
-        {
-            let batch_1 = state.create_batch(&system_context).unwrap();
-            assert!(state
-                .propose_commit_batch(CommitBatchArguments {
-                    batch_id: batch_1.clone(),
-                    operations: vec![BatchOperation::CreateAsset(CreateAssetArguments {
-                        key: "/d".to_string(),
-                        content_type: "text/plain".to_string(),
-                        max_age: Some(98),
-                        headers: Some(BTreeMap::from([
-                            ("H1".to_string(), "V1".to_string()),
-                            ("H2".to_string(), "V2".to_string())
-                        ])),
-                        enable_aliasing: Some(true),
-                        allow_raw_access: Some(false),
-                    }),],
-                })
-                .is_ok());
-            let evidence_1 = run_computation_until_completion(|_progress| {
-                state.compute_evidence(&ComputeEvidenceArguments {
-                    batch_id: batch_1.clone(),
-                    max_iterations: Some(3),
-                })
-            })
-            .unwrap();
-            delete_batch(&mut state, batch_1);
-
-            let batch_2 = state.create_batch(&system_context).unwrap();
-            assert!(state
-                .propose_commit_batch(CommitBatchArguments {
-                    batch_id: batch_2.clone(),
-                    operations: vec![BatchOperation::CreateAsset(CreateAssetArguments {
-                        key: "/d".to_string(),
-                        content_type: "text/plain".to_string(),
-                        max_age: Some(98),
-                        headers: Some(BTreeMap::from([
-                            ("H1".to_string(), "V1".to_string()),
-                            ("H2".to_string(), "V2".to_string())
-                        ])),
-                        enable_aliasing: Some(true),
-                        allow_raw_access: Some(false),
-                    }),],
-                })
-                .is_ok());
-            let evidence_2 = run_computation_until_completion(|_progress| {
-                state.compute_evidence(&ComputeEvidenceArguments {
-                    batch_id: batch_2.clone(),
-                    max_iterations: Some(3),
-                })
-            })
-            .unwrap();
-            assert_eq!(evidence_1, evidence_2);
-        }
-    }
-
-    #[test]
-    fn create_asset_arguments_key_affects_evidence() {
-        let mut state = State::default();
-        let system_context = mock_system_context();
-
-        let batch_1 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_1.clone(),
-                operations: vec![BatchOperation::CreateAsset(CreateAssetArguments {
-                    key: "/a/b/c".to_string(),
-                    content_type: "".to_string(),
-                    max_age: None,
-                    headers: None,
-                    enable_aliasing: None,
-                    allow_raw_access: None,
-                }),],
-            })
-            .is_ok());
-        let evidence_1 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_1.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-        delete_batch(&mut state, batch_1);
-
-        let batch_2 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_2.clone(),
-                operations: vec![BatchOperation::CreateAsset(CreateAssetArguments {
-                    key: "/d/e/f".to_string(),
-                    content_type: "".to_string(),
-                    max_age: None,
-                    headers: None,
-                    enable_aliasing: None,
-                    allow_raw_access: None,
-                }),],
-            })
-            .is_ok());
-        let evidence_2 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_2.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-
-        assert_ne!(evidence_1, evidence_2);
-    }
-
-    #[test]
-    fn create_asset_arguments_content_type_affects_evidence() {
-        let mut state = State::default();
-        let system_context = mock_system_context();
-
-        let batch_1 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_1.clone(),
-                operations: vec![BatchOperation::CreateAsset(CreateAssetArguments {
-                    key: "/".to_string(),
-                    content_type: "text/plain".to_string(),
-                    max_age: None,
-                    headers: None,
-                    enable_aliasing: None,
-                    allow_raw_access: None,
-                }),],
-            })
-            .is_ok());
-        let evidence_1 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_1.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-        delete_batch(&mut state, batch_1);
-
-        let batch_2 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_2.clone(),
-                operations: vec![BatchOperation::CreateAsset(CreateAssetArguments {
-                    key: "/".to_string(),
-                    content_type: "application/octet-stream".to_string(),
-                    max_age: None,
-                    headers: None,
-                    enable_aliasing: None,
-                    allow_raw_access: None,
-                }),],
-            })
-            .is_ok());
-        let evidence_2 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_2.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-
-        assert_ne!(evidence_1, evidence_2);
-    }
-
-    #[test]
-    fn create_asset_arguments_max_age_affects_evidence() {
-        let mut state = State::default();
-        let system_context = mock_system_context();
-
-        let batch_1 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_1.clone(),
-                operations: vec![BatchOperation::CreateAsset(CreateAssetArguments {
-                    key: "/".to_string(),
-                    content_type: "".to_string(),
-                    max_age: None,
-                    headers: None,
-                    enable_aliasing: None,
-                    allow_raw_access: None,
-                }),],
-            })
-            .is_ok());
-        let evidence_1 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_1.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-        delete_batch(&mut state, batch_1);
-
-        let batch_2 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_2.clone(),
-                operations: vec![BatchOperation::CreateAsset(CreateAssetArguments {
-                    key: "/".to_string(),
-                    content_type: "".to_string(),
-                    max_age: Some(32),
-                    headers: None,
-                    enable_aliasing: None,
-                    allow_raw_access: None,
-                }),],
-            })
-            .is_ok());
-
-        let evidence_2 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_2.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-        delete_batch(&mut state, batch_2);
-
-        let batch_3 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_3.clone(),
-                operations: vec![BatchOperation::CreateAsset(CreateAssetArguments {
-                    key: "/".to_string(),
-                    content_type: "".to_string(),
-                    max_age: Some(987),
-                    headers: None,
-                    enable_aliasing: None,
-                    allow_raw_access: None,
-                }),],
-            })
-            .is_ok());
-        let evidence_3 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_3.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-
-        assert_ne!(evidence_1, evidence_2);
-        assert_ne!(evidence_1, evidence_3);
-        assert_ne!(evidence_2, evidence_3);
-    }
-
-    #[test]
-    fn create_asset_arguments_headers_affect_evidence() {
-        let mut state = State::default();
-        let system_context = mock_system_context();
-
-        let batch_1 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_1.clone(),
-                operations: vec![BatchOperation::CreateAsset(CreateAssetArguments {
-                    key: "/".to_string(),
-                    content_type: "".to_string(),
-                    max_age: None,
-                    headers: Some(BTreeMap::from([("H1".to_string(), "V1".to_string()),])),
-                    enable_aliasing: None,
-                    allow_raw_access: None,
-                }),],
-            })
-            .is_ok());
-        let evidence_1 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_1.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-        delete_batch(&mut state, batch_1);
-
-        let batch_2 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_2.clone(),
-                operations: vec![BatchOperation::CreateAsset(CreateAssetArguments {
-                    key: "/".to_string(),
-                    content_type: "".to_string(),
-                    max_age: None,
-                    headers: Some(BTreeMap::from([("H1".to_string(), "V2".to_string()),])),
-                    enable_aliasing: None,
-                    allow_raw_access: None,
-                }),],
-            })
-            .is_ok());
-        let evidence_2 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_2.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-        delete_batch(&mut state, batch_2);
-
-        let batch_3 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_3.clone(),
-                operations: vec![BatchOperation::CreateAsset(CreateAssetArguments {
-                    key: "/".to_string(),
-                    content_type: "".to_string(),
-                    max_age: None,
-                    headers: Some(BTreeMap::from([("H2".to_string(), "V1".to_string()),])),
-                    enable_aliasing: None,
-                    allow_raw_access: None,
-                }),],
-            })
-            .is_ok());
-
-        let evidence_3 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_3.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-        delete_batch(&mut state, batch_3);
-
-        let batch_4 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_4.clone(),
-                operations: vec![BatchOperation::CreateAsset(CreateAssetArguments {
-                    key: "/".to_string(),
-                    content_type: "".to_string(),
-                    max_age: None,
-                    headers: Some(BTreeMap::from([
-                        ("H1".to_string(), "V1".to_string()),
-                        ("H2".to_string(), "V2".to_string()),
-                    ])),
-                    enable_aliasing: None,
-                    allow_raw_access: None,
-                }),],
-            })
-            .is_ok());
-        let evidence_4 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_4.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-
-        assert_ne!(evidence_1, evidence_2);
-        assert_ne!(evidence_1, evidence_3);
-        assert_ne!(evidence_1, evidence_4);
-        assert_ne!(evidence_2, evidence_3);
-        assert_ne!(evidence_2, evidence_4);
-        assert_ne!(evidence_3, evidence_4);
-    }
-
-    #[test]
-    fn create_asset_arguments_enable_aliasing_affects_evidence() {
-        let mut state = State::default();
-        let system_context = mock_system_context();
-
-        let batch_1 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_1.clone(),
-                operations: vec![BatchOperation::CreateAsset(CreateAssetArguments {
-                    key: "/".to_string(),
-                    content_type: "".to_string(),
-                    max_age: None,
-                    headers: None,
-                    enable_aliasing: None,
-                    allow_raw_access: None,
-                }),],
-            })
-            .is_ok());
-        let evidence_1 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_1.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-        delete_batch(&mut state, batch_1);
-
-        let batch_2 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_2.clone(),
-                operations: vec![BatchOperation::CreateAsset(CreateAssetArguments {
-                    key: "/".to_string(),
-                    content_type: "".to_string(),
-                    max_age: None,
-                    headers: None,
-                    enable_aliasing: Some(false),
-                    allow_raw_access: None,
-                }),],
-            })
-            .is_ok());
-
-        let evidence_2 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_2.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-        delete_batch(&mut state, batch_2);
-
-        let batch_3 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_3.clone(),
-                operations: vec![BatchOperation::CreateAsset(CreateAssetArguments {
-                    key: "/".to_string(),
-                    content_type: "".to_string(),
-                    max_age: None,
-                    headers: None,
-                    enable_aliasing: Some(true),
-                    allow_raw_access: None,
-                }),],
-            })
-            .is_ok());
-        let evidence_3 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_3.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-
-        assert_ne!(evidence_1, evidence_2);
-        assert_ne!(evidence_1, evidence_3);
-        assert_ne!(evidence_2, evidence_3);
-    }
-
-    #[test]
-    fn create_asset_arguments_allow_raw_access_affects_evidence() {
-        let mut state = State::default();
-        let system_context = mock_system_context();
-
-        let batch_1 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_1.clone(),
-                operations: vec![BatchOperation::CreateAsset(CreateAssetArguments {
-                    key: "/".to_string(),
-                    content_type: "".to_string(),
-                    max_age: None,
-                    headers: None,
-                    enable_aliasing: None,
-                    allow_raw_access: None,
-                }),],
-            })
-            .is_ok());
-        let evidence_1 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_1.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-        delete_batch(&mut state, batch_1);
-
-        let batch_2 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_2.clone(),
-                operations: vec![BatchOperation::CreateAsset(CreateAssetArguments {
-                    key: "/".to_string(),
-                    content_type: "".to_string(),
-                    max_age: None,
-                    headers: None,
-                    enable_aliasing: None,
-                    allow_raw_access: Some(false),
-                }),],
-            })
-            .is_ok());
-
-        let evidence_2 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_2.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-        delete_batch(&mut state, batch_2);
-
-        let batch_3 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_3.clone(),
-                operations: vec![BatchOperation::CreateAsset(CreateAssetArguments {
-                    key: "/".to_string(),
-                    content_type: "".to_string(),
-                    max_age: None,
-                    headers: None,
-                    enable_aliasing: None,
-                    allow_raw_access: Some(true),
-                }),],
-            })
-            .is_ok());
-        let evidence_3 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_3.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-
-        assert_ne!(evidence_1, evidence_2);
-        assert_ne!(evidence_1, evidence_3);
-        assert_ne!(evidence_2, evidence_3);
-    }
-
-    #[test]
-    fn set_asset_content_arguments_key_affects_evidence() {
-        let mut state = State::default();
-        let system_context = mock_system_context();
-
-        let batch_1 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_1.clone(),
-                operations: vec![SetAssetContent(SetAssetContentArguments {
-                    key: "/1".to_string(),
-                    content_encoding: "identity".to_string(),
-                    chunk_ids: vec![],
-                    last_chunk: None,
-                    sha256: None,
-                })],
-            })
-            .is_ok());
-        let evidence_1 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_1.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-        delete_batch(&mut state, batch_1);
-
-        let batch_2 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_2.clone(),
-                operations: vec![SetAssetContent(SetAssetContentArguments {
-                    key: "/2".to_string(),
-                    content_encoding: "identity".to_string(),
-                    chunk_ids: vec![],
-                    last_chunk: None,
-                    sha256: None,
-                })],
-            })
-            .is_ok());
-        let evidence_2 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_2.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-
-        assert_ne!(evidence_1, evidence_2);
-    }
-
-    #[test]
-    fn set_asset_content_arguments_content_encoding_affects_evidence() {
-        let mut state = State::default();
-        let system_context = mock_system_context();
-
-        let batch_1 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_1.clone(),
-                operations: vec![SetAssetContent(SetAssetContentArguments {
-                    key: "/1".to_string(),
-                    content_encoding: "identity".to_string(),
-                    chunk_ids: vec![],
-                    last_chunk: None,
-                    sha256: None,
-                })],
-            })
-            .is_ok());
-        let evidence_1 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_1.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-        delete_batch(&mut state, batch_1);
-
-        let batch_2 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_2.clone(),
-                operations: vec![SetAssetContent(SetAssetContentArguments {
-                    key: "/1".to_string(),
-                    content_encoding: "gzip".to_string(),
-                    chunk_ids: vec![],
-                    last_chunk: None,
-                    sha256: None,
-                })],
-            })
-            .is_ok());
-        let evidence_2 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_2.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-
-        assert_ne!(evidence_1, evidence_2);
-    }
-
-    #[test]
-    fn set_asset_content_arguments_chunk_contents_affects_evidence() {
-        let mut state = State::default();
-        let system_context = mock_system_context();
-
-        const CHUNK_1_CONTENT: &[u8] = b"first batch chunk content";
-        const CHUNK_2_CONTENT: &[u8] = b"second batch chunk content";
-
-        let batch_1 = state.create_batch(&system_context).unwrap();
-        let chunk_1 = state
-            .create_chunk(
-                CreateChunkArg {
-                    batch_id: batch_1.clone(),
-                    content: ByteBuf::from(CHUNK_1_CONTENT),
-                },
-                &system_context,
-            )
-            .unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_1.clone(),
-                operations: vec![SetAssetContent(SetAssetContentArguments {
-                    key: "/1".to_string(),
-                    content_encoding: "identity".to_string(),
-                    chunk_ids: vec![chunk_1],
-                    last_chunk: None,
-                    sha256: None,
-                })],
-            })
-            .is_ok());
-        let evidence_1 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_1.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-        delete_batch(&mut state, batch_1);
-
-        let batch_2 = state.create_batch(&system_context).unwrap();
-        let chunk_2 = state
-            .create_chunk(
-                CreateChunkArg {
-                    batch_id: batch_2.clone(),
-                    content: ByteBuf::from(CHUNK_2_CONTENT),
-                },
-                &system_context,
-            )
-            .unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_2.clone(),
-                operations: vec![SetAssetContent(SetAssetContentArguments {
-                    key: "/1".to_string(),
-                    content_encoding: "identity".to_string(),
-                    chunk_ids: vec![chunk_2],
-                    last_chunk: None,
-                    sha256: None,
-                })],
-            })
-            .is_ok());
-        let evidence_2 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_2.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-
-        assert_ne!(evidence_1, evidence_2);
-    }
-    #[test]
-    fn set_asset_content_arguments_multiple_chunk_contents_affects_evidence() {
-        let mut state = State::default();
-        let system_context = mock_system_context();
-
-        const CHUNK_1_CONTENT: &[u8] = b"first chunk, same for both";
-        const BATCH_1_CHUNK_2_CONTENT: &[u8] = b"first batch second chunk content";
-        const BATCH_2_CHUNK_2_CONTENT: &[u8] = b"second batch second chunk content";
-
-        let batch_1 = state.create_batch(&system_context).unwrap();
-        {
-            let chunk_1 = state
-                .create_chunk(
-                    CreateChunkArg {
-                        batch_id: batch_1.clone(),
-                        content: ByteBuf::from(CHUNK_1_CONTENT),
-                    },
-                    &system_context,
-                )
-                .unwrap();
-            let chunk_2 = state
-                .create_chunk(
-                    CreateChunkArg {
-                        batch_id: batch_1.clone(),
-                        content: ByteBuf::from(BATCH_1_CHUNK_2_CONTENT),
-                    },
-                    &system_context,
-                )
-                .unwrap();
-
-            assert!(state
-                .propose_commit_batch(CommitBatchArguments {
-                    batch_id: batch_1.clone(),
-                    operations: vec![SetAssetContent(SetAssetContentArguments {
-                        key: "/1".to_string(),
-                        content_encoding: "identity".to_string(),
-                        chunk_ids: vec![chunk_1, chunk_2],
-                        last_chunk: None,
-                        sha256: None,
-                    })],
-                })
-                .is_ok());
-        }
-        let evidence_1 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_1.clone(),
-                max_iterations: Some(4),
-            })
-        })
-        .unwrap();
-        delete_batch(&mut state, batch_1);
-
-        let batch_2 = state.create_batch(&system_context).unwrap();
-        {
-            let chunk_1 = state
-                .create_chunk(
-                    CreateChunkArg {
-                        batch_id: batch_2.clone(),
-                        content: ByteBuf::from(CHUNK_1_CONTENT),
-                    },
-                    &system_context,
-                )
-                .unwrap();
-            let chunk_2 = state
-                .create_chunk(
-                    CreateChunkArg {
-                        batch_id: batch_2.clone(),
-                        content: ByteBuf::from(BATCH_2_CHUNK_2_CONTENT),
-                    },
-                    &system_context,
-                )
-                .unwrap();
-            assert!(state
-                .propose_commit_batch(CommitBatchArguments {
-                    batch_id: batch_2.clone(),
-                    operations: vec![SetAssetContent(SetAssetContentArguments {
-                        key: "/1".to_string(),
-                        content_encoding: "identity".to_string(),
-                        chunk_ids: vec![chunk_1, chunk_2],
-                        last_chunk: None,
-                        sha256: None,
-                    })],
-                })
-                .is_ok());
-        }
-        let evidence_2 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_2.clone(),
-                max_iterations: Some(4),
-            })
-        })
-        .unwrap();
-
-        assert_ne!(evidence_1, evidence_2);
-    }
-
-    #[test]
-    fn set_asset_content_arguments_sha256_affects_evidence() {
-        let mut state = State::default();
-        let system_context = mock_system_context();
-
-        let sha256_1 = ByteBuf::from("01020304");
-        let sha256_2 = ByteBuf::from("09080706");
-
-        let batch_1 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_1.clone(),
-                operations: vec![SetAssetContent(SetAssetContentArguments {
-                    key: "/1".to_string(),
-                    content_encoding: "identity".to_string(),
-                    last_chunk: None,
-                    chunk_ids: vec![],
-                    sha256: Some(sha256_1),
-                })],
-            })
-            .is_ok());
-        let evidence_1 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_1.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-        delete_batch(&mut state, batch_1);
-
-        let batch_2 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_2.clone(),
-                operations: vec![SetAssetContent(SetAssetContentArguments {
-                    key: "/1".to_string(),
-                    content_encoding: "identity".to_string(),
-                    last_chunk: None,
-                    chunk_ids: vec![],
-                    sha256: Some(sha256_2),
-                })],
-            })
-            .is_ok());
-        let evidence_2 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_2.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-
-        assert_ne!(evidence_1, evidence_2);
-    }
-
-    #[test]
-    fn unset_asset_content_arguments_key_affects_evidence() {
-        let mut state = State::default();
-        let system_context = mock_system_context();
-
-        let batch_1 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_1.clone(),
-                operations: vec![BatchOperation::UnsetAssetContent(
-                    UnsetAssetContentArguments {
-                        key: "/1".to_string(),
-                        content_encoding: "".to_string(),
-                    }
-                ),],
-            })
-            .is_ok());
-        let evidence_1 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_1.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-        delete_batch(&mut state, batch_1);
-
-        let batch_2 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_2.clone(),
-                operations: vec![BatchOperation::UnsetAssetContent(
-                    UnsetAssetContentArguments {
-                        key: "/2".to_string(),
-                        content_encoding: "".to_string(),
-                    }
-                ),],
-            })
-            .is_ok());
-        let evidence_2 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_2.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-
-        assert_ne!(evidence_1, evidence_2);
-    }
-
-    #[test]
-    fn unset_asset_content_arguments_content_encoding_affects_evidence() {
-        let mut state = State::default();
-        let system_context = mock_system_context();
-
-        let batch_1 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_1.clone(),
-                operations: vec![BatchOperation::UnsetAssetContent(
-                    UnsetAssetContentArguments {
-                        key: "/1".to_string(),
-                        content_encoding: "identity".to_string(),
-                    }
-                ),],
-            })
-            .is_ok());
-        let evidence_1 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_1.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-        delete_batch(&mut state, batch_1);
-
-        let batch_2 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_2.clone(),
-                operations: vec![BatchOperation::UnsetAssetContent(
-                    UnsetAssetContentArguments {
-                        key: "/1".to_string(),
-                        content_encoding: "gzip".to_string(),
-                    }
-                ),],
-            })
-            .is_ok());
-        let evidence_2 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_2.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-
-        assert_ne!(evidence_1, evidence_2);
-    }
-
-    #[test]
-    fn delete_asset_content_arguments_key_affects_evidence() {
-        // todo
-        let mut state = State::default();
-        let system_context = mock_system_context();
-
-        let batch_1 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_1.clone(),
-                operations: vec![BatchOperation::DeleteAsset(DeleteAssetArguments {
-                    key: "/1".to_string(),
-                }),],
-            })
-            .is_ok());
-        let evidence_1 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_1.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-        delete_batch(&mut state, batch_1);
-
-        let batch_2 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_2.clone(),
-                operations: vec![BatchOperation::DeleteAsset(DeleteAssetArguments {
-                    key: "/2".to_string(),
-                }),],
-            })
-            .is_ok());
-        let evidence_2 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_2.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-
-        assert_ne!(evidence_1, evidence_2);
-    }
-
-    #[test]
-    fn clear_affects_evidence() {
-        let mut state = State::default();
-        let system_context = mock_system_context();
-
-        let batch_1 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_1.clone(),
-                operations: vec![BatchOperation::Clear(ClearArguments {}),],
-            })
-            .is_ok());
-        let evidence_1 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_1.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-        delete_batch(&mut state, batch_1);
-
-        let batch_2 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_2.clone(),
-                operations: vec![
-                    BatchOperation::Clear(ClearArguments {}),
-                    BatchOperation::Clear(ClearArguments {})
-                ],
-            })
-            .is_ok());
-        let evidence_2 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_2.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-
-        assert_ne!(evidence_1, evidence_2);
-    }
-
-    #[test]
-    fn set_asset_properties_arguments_key_affects_evidence() {
-        let mut state = State::default();
-        let system_context = mock_system_context();
-
-        let batch_1 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_1.clone(),
-                operations: vec![BatchOperation::SetAssetProperties(
-                    SetAssetPropertiesArguments {
-                        key: "/1".to_string(),
-                        max_age: Some(Some(100)),
-                        headers: None,
-                        allow_raw_access: Some(Some(false)),
-                        is_aliased: Some(Some(true))
-                    }
-                ),],
-            })
-            .is_ok());
-        let evidence_1 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_1.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-        delete_batch(&mut state, batch_1);
-
-        let batch_2 = state.create_batch(&system_context).unwrap();
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_2.clone(),
-                operations: vec![BatchOperation::SetAssetProperties(
-                    SetAssetPropertiesArguments {
-                        key: "/2".to_string(),
-                        max_age: Some(Some(100)),
-                        headers: None,
-                        allow_raw_access: Some(Some(false)),
-                        is_aliased: Some(Some(true))
-                    }
-                ),],
-            })
-            .is_ok());
-        let evidence_2 = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_2.clone(),
-                max_iterations: Some(3),
-            })
-        })
-        .unwrap();
-
-        assert_ne!(evidence_1, evidence_2);
-    }
-
-    #[test]
-    fn set_asset_properties_arguments_properties_affects_evidence() {
-        fn generate_unique_set_asset_properties() -> Vec<SetAssetPropertiesArguments> {
-            let mut result = vec![];
-            for max_age in &[None, Some(None), Some(Some(100))] {
-                for headers in &[
-                    None,
-                    Some(None),
-                    Some(Some(BTreeMap::from([(
-                        String::from("a"),
-                        String::from("b"),
-                    )]))),
-                ] {
-                    for allow_raw_access in &[None, Some(None), Some(Some(true)), Some(Some(false))]
-                    {
-                        for is_aliased in &[None, Some(None), Some(Some(true)), Some(Some(false))] {
-                            result.push(SetAssetPropertiesArguments {
-                                key: "/1".to_string(),
-                                max_age: *max_age,
-                                headers: headers.clone(),
-                                allow_raw_access: *allow_raw_access,
-                                is_aliased: *is_aliased,
-                            });
-                        }
-                    }
-                }
-            }
-            result
-        }
-
-        fn compute_evidence_for_set_asset_properties(
-            args: SetAssetPropertiesArguments,
-        ) -> serde_bytes::ByteBuf {
-            let mut state = State::default();
-            let system_context = mock_system_context();
-
-            let batch = state.create_batch(&system_context).unwrap();
-            assert!(state
-                .propose_commit_batch(CommitBatchArguments {
-                    batch_id: batch.clone(),
-                    operations: vec![BatchOperation::SetAssetProperties(args)],
-                })
-                .is_ok());
-
-            run_computation_until_completion(|_progress| {
-                state.compute_evidence(&ComputeEvidenceArguments {
-                    batch_id: batch.clone(),
-                    max_iterations: Some(3),
-                })
-            })
-            .unwrap()
-        }
-
-        let instances = generate_unique_set_asset_properties();
-        let evidences = instances
-            .into_iter()
-            .map(compute_evidence_for_set_asset_properties)
-            .collect::<Vec<_>>();
-
-        // Check if all evidences are different.
-        for i in 0..evidences.len() {
-            for j in (i + 1)..evidences.len() {
-                assert_ne!(evidences[i], evidences[j]);
-            }
-        }
-    }
-}
-
-#[cfg(test)]
-mod validate_commit_proposed_batch {
-    use super::*;
-    use crate::types::ComputeEvidenceArguments;
-
-    #[test]
-    fn batch_not_found() {
-        let mut state = State::default();
-        let system_context = mock_system_context();
-
-        match state.validate_commit_proposed_batch(CommitProposedBatchArguments {
-            batch_id: 1_u8.into(),
-            evidence: Default::default(),
-        }) {
-            Err(err) if err.contains("batch not found") => (),
-            other => panic!("expected 'batch not found' error, got: {other:?}"),
-        }
-
-        match run_computation_until_completion(|progress| {
-            state.commit_proposed_batch(
-                &CommitProposedBatchArguments {
-                    batch_id: 1_u8.into(),
-                    evidence: Default::default(),
-                },
-                progress,
-                &system_context,
-            )
-        }) {
-            Err(err) if err.contains("batch not found") => (),
-            other => panic!("expected 'batch not found' error, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn no_commit_batch_arguments() {
-        let mut state = State::default();
-        let system_context = mock_system_context();
-        let batch_id = state.create_batch(&system_context).unwrap();
-
-        match state.validate_commit_proposed_batch(CommitProposedBatchArguments {
-            batch_id: batch_id.clone(),
-            evidence: Default::default(),
-        }) {
-            Err(err) if err.contains("batch does not have CommitBatchArguments") => (),
-            other => panic!("expected 'batch not found' error, got: {other:?}"),
-        }
-
-        match run_computation_until_completion(|progress| {
-            state.commit_proposed_batch(
-                &CommitProposedBatchArguments {
-                    batch_id: batch_id.clone(),
-                    evidence: Default::default(),
-                },
-                progress,
-                &system_context,
-            )
-        }) {
-            Err(err) if err.contains("batch does not have CommitBatchArguments") => (),
-            other => panic!("expected 'batch not found' error, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn evidence_not_computed() {
-        let mut state = State::default();
-        let system_context = mock_system_context();
-        let batch_id = state.create_batch(&system_context).unwrap();
-
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_id.clone(),
-                operations: vec![],
-            })
-            .is_ok());
-
-        match state.validate_commit_proposed_batch(CommitProposedBatchArguments {
-            batch_id: batch_id.clone(),
-            evidence: Default::default(),
-        }) {
-            Err(err) if err.contains("batch does not have computed evidence") => (),
-            other => panic!("expected 'batch not found' error, got: {other:?}"),
-        }
-        match run_computation_until_completion(|progress| {
-            state.commit_proposed_batch(
-                &CommitProposedBatchArguments {
-                    batch_id: batch_id.clone(),
-                    evidence: Default::default(),
-                },
-                progress,
-                &system_context,
-            )
-        }) {
-            Err(err) if err.contains("batch does not have computed evidence") => (),
-            other => panic!("expected 'batch not found' error, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn evidence_does_not_match() {
-        let mut state = State::default();
-        let system_context = mock_system_context();
-        let batch_id = state.create_batch(&system_context).unwrap();
-
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_id.clone(),
-                operations: vec![],
-            })
-            .is_ok());
-
-        assert!(run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_id.clone(),
-                max_iterations: Some(1),
-            })
-        })
-        .is_ok());
-
-        match state.validate_commit_proposed_batch(CommitProposedBatchArguments {
-            batch_id: batch_id.clone(),
-            evidence: Default::default(),
-        }) {
-            Err(err) if err.contains("does not match presented evidence") => (),
-            other => panic!("expected 'batch not found' error, got: {other:?}"),
-        }
-
-        match run_computation_until_completion(|progress| {
-            state.commit_proposed_batch(
-                &CommitProposedBatchArguments {
-                    batch_id: batch_id.clone(),
-                    evidence: Default::default(),
-                },
-                progress,
-                &system_context,
-            )
-        }) {
-            Err(err) if err.contains("does not match presented evidence") => (),
-            other => panic!("expected 'batch not found' error, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn all_good() {
-        let mut state = State::default();
-        let system_context = mock_system_context();
-        let batch_id = state.create_batch(&system_context).unwrap();
-
-        assert!(state
-            .propose_commit_batch(CommitBatchArguments {
-                batch_id: batch_id.clone(),
-                operations: vec![],
-            })
-            .is_ok());
-
-        let compute_evidence_result = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_id.clone(),
-                max_iterations: Some(1),
-            })
-        });
-        assert!(compute_evidence_result.is_ok());
-
-        let evidence = if let Ok(computed_evidence) = compute_evidence_result {
-            computed_evidence
-        } else {
-            unreachable!()
-        };
-
-        assert_eq!(
-            state
-                .validate_commit_proposed_batch(CommitProposedBatchArguments {
-                    batch_id: batch_id.clone(),
-                    evidence: evidence.clone(),
-                },)
-                .unwrap(),
-            "commit proposed batch 0 with evidence e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
-        );
-
-        run_computation_until_completion(|progress| {
-            state.commit_proposed_batch(
-                &CommitProposedBatchArguments {
-                    batch_id: batch_id.clone(),
-                    evidence: evidence.clone(),
-                },
-                progress,
-                &system_context,
-            )
-        })
-        .unwrap();
-    }
-}
-
-#[cfg(test)]
 mod configuration_methods {
     use super::*;
     use crate::types::ConfigureArguments;
@@ -4255,56 +2384,6 @@ mod configuration_methods {
 mod enforce_limits {
     use super::*;
     use crate::types::ConfigureArguments;
-
-    #[test]
-    fn cannot_create_batch_if_batch_already_proposed_with_no_batch_limit() {
-        cannot_create_batch_if_batch_already_proposed_with_batch_limit(None);
-    }
-
-    #[test]
-    fn cannot_create_batch_if_batch_already_proposed_with_a_batch_limit() {
-        // test with a batch limit to make sure we get the right message (not: batch limit exceeded)
-        cannot_create_batch_if_batch_already_proposed_with_batch_limit(Some(1));
-    }
-
-    fn cannot_create_batch_if_batch_already_proposed_with_batch_limit(max_batches: Option<u64>) {
-        let mut state = State::default();
-        let mut system_context = mock_system_context();
-        let time_now = system_context.current_timestamp_ns;
-
-        state.configure(ConfigureArguments {
-            max_batches: Some(max_batches),
-            max_chunks: None,
-            max_bytes: None,
-        });
-
-        let batch_id = state.create_batch(&system_context).unwrap();
-        let cba = CommitBatchArguments {
-            batch_id: batch_id.clone(),
-            operations: vec![],
-        };
-        assert!(state.propose_commit_batch(cba).is_ok());
-
-        system_context.current_timestamp_ns = time_now + BATCH_EXPIRY_NANOS - 1;
-        assert_eq!(
-            state.create_batch(&system_context).unwrap_err(),
-            "Batch 0 has not completed evidence computation.  Wait for it to expire or delete it to propose another."
-        );
-
-        assert!(run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_id.clone(),
-                max_iterations: Some(1),
-            })
-        })
-        .is_ok());
-
-        system_context.current_timestamp_ns = time_now + BATCH_EXPIRY_NANOS + 1;
-        assert_eq!(
-            state.create_batch(&system_context).unwrap_err(),
-            "Batch 0 is already proposed.  Delete or execute it to propose another."
-        );
-    }
 
     #[test]
     fn max_batches() {
@@ -5121,78 +3200,6 @@ mod set_asset_content_sha256_verification {
 #[cfg(test)]
 mod compute_state_hash {
     use super::*;
-    use sha2::{Digest, Sha256};
-
-    #[test]
-    fn test_compute_state_hash_matches_evidence() {
-        let mut state = State::default();
-        let system_context = mock_system_context();
-
-        // Create a batch to compute evidence "normally"
-        let batch_id = state.create_batch(&system_context).unwrap();
-
-        let chunk_id = state
-            .create_chunk(
-                CreateChunkArg {
-                    batch_id: batch_id.clone(),
-                    content: ByteBuf::from(b"content1"),
-                },
-                &system_context,
-            )
-            .unwrap();
-
-        // Compute SHA256 of content
-        let mut hasher = Sha256::new();
-        hasher.update(b"content1");
-        let sha256: [u8; 32] = hasher.finalize().into();
-
-        let args = CommitBatchArguments {
-            batch_id: batch_id.clone(),
-            operations: vec![
-                BatchOperation::CreateAsset(CreateAssetArguments {
-                    key: "asset1".to_string(),
-                    content_type: "text/plain".to_string(),
-                    max_age: None,
-                    headers: None,
-                    enable_aliasing: None,
-                    allow_raw_access: None,
-                }),
-                BatchOperation::SetAssetContent(SetAssetContentArguments {
-                    key: "asset1".to_string(),
-                    content_encoding: "identity".to_string(),
-                    chunk_ids: vec![chunk_id],
-                    last_chunk: None,
-                    sha256: Some(ByteBuf::from(sha256)),
-                }),
-            ],
-        };
-
-        state.propose_commit_batch(args.clone()).unwrap();
-
-        let evidence = run_computation_until_completion(|_progress| {
-            state.compute_evidence(&ComputeEvidenceArguments {
-                batch_id: batch_id.clone(),
-                max_iterations: None,
-            })
-        })
-        .unwrap();
-
-        // Now apply the batch to state so we can compute state hash
-        run_computation_until_completion(|progress| {
-            state.commit_batch(&args, progress, &system_context)
-        })
-        .unwrap();
-
-        // Compute state hash
-        let state_hash =
-            run_computation_until_completion(|_progress| state.compute_state_hash()).unwrap();
-
-        assert_eq!(
-            hex::encode(evidence.as_slice()),
-            state_hash,
-            "State hash should match evidence computed from batch when starting with an empty asset canister"
-        );
-    }
 
     #[test]
     fn test_compute_state_hash_interruption() {
