+++ ic-certified-assets/src/types.rs
@@ -81,20 +81,8 @@ pub struct CommitBatchArguments {
 }
 
 #[derive(Clone, Debug, CandidType, Deserialize)]
-pub struct CommitProposedBatchArguments {
-    pub batch_id: BatchId,
-    pub evidence: ByteBuf,
-}
-
-#[derive(Clone, Debug, CandidType, Deserialize)]
 pub struct DeleteBatchArguments {
     pub batch_id: BatchId,
-}
-
-#[derive(Clone, Debug, CandidType, Deserialize)]
-pub struct ComputeEvidenceArguments {
-    pub batch_id: BatchId,
-    pub max_iterations: Option<u16>,
 }
 
 #[derive(Clone, Debug, CandidType, Deserialize)]
